Improvements to code
Major improvements to the code, including removal of the javascript part
and replacing it with pure css solution. Added transitions to shop,
buttons and several others main parts of the website. Fixed the problems
with the position of the bicycle logos on the images in the work section
of the website. And some personal changes to make the website look
better. More soon.